# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/entites": {
      "get": {
        "description": "Les entités privées en sont absentes, sans mention — un décompte manquant révélerait leur existence.",
        "operationId": "EntitesController_lister",
        "parameters": [
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "etat",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "actif",
                "archive"
              ],
              "type": "string"
            }
          },
          {
            "name": "decalage",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "limite",
            "required": false,
            "in": "query",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EntiteResumeeDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "jeton": []
          }
        ],
        "summary": "Annuaire filtrable",
        "tags": [
          "entites"
        ]
      },
      "post": {
        "description": "Source, fiabilité et date de constatation données au niveau de la requête servent de valeurs par défaut à chaque fait — c’est le bandeau de source active.",
        "operationId": "EntitesController_creer",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreationEntiteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FicheEntiteDto"
                }
              }
            }
          },
          "409": {
            "description": "Valeur unique déjà attribuée"
          }
        },
        "security": [
          {
            "jeton": []
          }
        ],
        "summary": "Création d’une entité et de ses premiers faits",
        "tags": [
          "entites"
        ]
      }
    },
    "/entites/similaires": {
      "get": {
        "description": "Similarité trigramme du libellé, et identité exacte d’une valeur unique du type. Ne propose jamais une entité privée.",
        "operationId": "EntitesController_similaires",
        "parameters": [
          {
            "name": "q",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SuggestionDoublonDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "jeton": []
          }
        ],
        "summary": "Détection de doublons à la frappe",
        "tags": [
          "entites"
        ]
      }
    },
    "/entites/{id}": {
      "get": {
        "description": "Champs projetés depuis les seuls faits visibles par l’agent, et liens lus depuis cette fiche — un lien est une arête unique, vue des deux côtés.",
        "operationId": "EntitesController_lire",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FicheEntiteDto"
                }
              }
            }
          },
          "404": {
            "description": "Inconnue, ou privée sans habilitation — jamais 403, qui confirmerait son existence"
          }
        },
        "security": [
          {
            "jeton": []
          }
        ],
        "summary": "Fiche assemblée",
        "tags": [
          "entites"
        ]
      },
      "patch": {
        "operationId": "EntitesController_modifier",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModificationEntiteDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FicheEntiteDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "jeton": []
          }
        ],
        "tags": [
          "entites"
        ]
      }
    },
    "/entites/{id}/historique": {
      "get": {
        "description": "Faits infirmés ou archivés — jamais supprimés, toujours consultables — et traces d’écriture du journal d’audit.",
        "operationId": "EntitesController_historique",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EvenementHistoriqueDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "jeton": []
          }
        ],
        "summary": "Onglet Historique",
        "tags": [
          "entites"
        ]
      }
    },
    "/entites/{id}/annuler-creation": {
      "post": {
        "description": "Retire une entité que le sous-formulaire venait de persister et que l’agent abandonne. Réservée à son auteur, dans l’heure, sur une entité que rien d’autre ne désigne — au-delà, seul l’archivage sort.",
        "operationId": "EntitesController_annulerCreation",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "409": {
            "description": "Trop ancienne, déjà référencée, ou saisie par un autre agent"
          }
        },
        "security": [
          {
            "jeton": []
          }
        ],
        "summary": "Annulation d’une saisie en cascade",
        "tags": [
          "entites"
        ]
      }
    },
    "/entites/{id}/fusion": {
      "post": {
        "description": "L’entité de l’URL est absorbée par celle du corps, qui reçoit ses faits, ses fichiers, ses suivis et ses habilitations. L’absorbée reste en base, archivée, et **redirige** : un ancien lien vers elle continue de mener quelque part. La fiche renvoyée est celle qui subsiste.",
        "operationId": "EntitesController_fusionner",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FusionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FicheEntiteDto"
                }
              }
            }
          },
          "409": {
            "description": "Types différents, ou entité déjà fusionnée"
          }
        },
        "security": [
          {
            "jeton": []
          }
        ],
        "summary": "Fusion de doublons",
        "tags": [
          "entites"
        ]
      }
    },
    "/entites/{id}/archiver": {
      "post": {
        "description": "Rien n’est jamais supprimé : l’entité sort des écrans courants et reste consultable, ses faits intacts.",
        "operationId": "EntitesController_archiver",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FicheEntiteDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "jeton": []
          }
        ],
        "summary": "Archivage",
        "tags": [
          "entites"
        ]
      }
    },
    "/entites/{id}/desarchiver": {
      "post": {
        "operationId": "EntitesController_desarchiver",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FicheEntiteDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "jeton": []
          }
        ],
        "tags": [
          "entites"
        ]
      }
    },
...
  },
  "info": {
    "title": "Centrale N&I",
    "description": "API du service Narcotics & Investigations de la SAHP. Le référentiel — types d'entités, champs, types de liens, onglets — est dynamique et n'apparaît pas dans ce contrat : il se récupère par GET /referentiel.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "jeton": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "EntiteResumeeDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "typeEntiteId": {
            "type": "string",
            "format": "uuid"
          },
          "typeCode": {
            "type": "string"
          },
          "libelle": {
            "type": "string"
          },
          "visibilite": {
            "enum": [
              "public",
              "restreint",
              "prive"
            ],
            "type": "string"
          },
          "etat": {
            "enum": [
              "actif",
              "archive"
            ],
            "type": "string"
          },
          "modifieLe": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "typeEntiteId",
          "typeCode",
          "libelle",
          "visibilite",
          "etat",
          "modifieLe"
        ]
      },
      "SuggestionDoublonDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "libelle": {
            "type": "string"
          },
          "typeCode": {
            "type": "string"
          },
          "proximite": {
            "type": "number",
            "description": "Similarité trigramme du libellé, entre 0 et 1"
          },
          "valeurUniqueIdentique": {
            "type": "boolean",
            "description": "Une valeur unique du type est identique — c’est un doublon sûr"
          }
        },
        "required": [
          "id",
          "libelle",
          "typeCode",
          "proximite",
          "valeurUniqueIdentique"
        ]
      },
      "RattachementDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "nom": {
            "type": "string"
          },
          "visibilite": {
            "enum": [
              "public",
              "restreint",
              "prive"
            ],
            "type": "string"
          },
          "estPivot": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "nom",
          "visibilite",
          "estPivot"
        ]
      },
...
      "ChampDeFicheDto": {
        "type": "object",
        "properties": {
          "definitionChampId": {
            "type": "string",
            "format": "uuid"
          },
          "cle": {
            "type": "string"
          },
          "libelle": {
            "type": "string"
          },
          "typeDonnee": {
            "enum": [
              "texte",
              "nombre",
              "date",
              "datetime",
              "booleen",
              "liste",
              "fichier"
            ],
            "type": "string"
          },
          "multiple": {
            "type": "boolean"
          },
          "valeur": {
            "nullable": true,
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "number"
              },
              {
                "type": "boolean"
              },
              {
                "type": "array",
                "items": {}
              }
            ],
            "description": "Valeur projetée, celle qui s’affiche en évidence"
          },
          "faits": {
            "description": "Les faits qui la soutiennent. Plus d’un signale un recoupement de sources.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FaitDeChampDto"
            }
          },
          "multiSources": {
            "type": "boolean",
            "description": "Plusieurs sources distinctes affirment la même valeur"
          }
        },
        "required": [
          "definitionChampId",
          "cle",
          "libelle",
          "typeDonnee",
          "multiple",
          "valeur",
          "faits",
          "multiSources"
        ]
      },
      "ExtremiteDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "libelle": {
            "type": "string"
          },
          "typeCode": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "libelle",
          "typeCode"
        ]
      },
      "LienDeFicheDto": {
        "type": "object",
        "properties": {
          "faitId": {
            "type": "string",
            "format": "uuid",
            "description": "Identifiant du fait"
          },
          "sens": {
            "enum": [
              "direct",
              "inverse"
            ],
            "type": "string",
            "description": "Le même fait, vu depuis l’une ou l’autre extrémité. Une seule arête existe en base."
          },
          "typeLienId": {
            "type": "string",
            "format": "uuid"
          },
          "libelle": {
            "type": "string",
            "description": "Libellé lu depuis cette fiche"
          },
          "autreEntite": {
            "$ref": "#/components/schemas/ExtremiteDto"
          },
          "source": {
            "type": "string"
          },
          "fiabilite": {
            "type": "number"
          },
          "dateConstatation": {
            "type": "string",
            "format": "date"
          },
          "visibilite": {
            "enum": [
              "public",
              "restreint",
              "prive"
            ],
            "type": "string"
          },
          "visibiliteEffective": {
            "enum": [
              "public",
              "restreint",
              "prive"
            ],
            "type": "string"
          }
        },
        "required": [
          "faitId",
          "sens",
          "typeLienId",
          "libelle",
          "autreEntite",
          "source",
          "fiabilite",
          "dateConstatation",
          "visibilite",
          "visibiliteEffective"
        ]
      },
      "OngletPeupleDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "libelle": {
            "type": "string",
            "example": "Membres"
          },
          "ordre": {
            "type": "number"
          },
          "compteur": {
            "type": "number",
            "description": "Ne compte que les liens visibles par cet agent — un compteur exhaustif révélerait ce qui est masqué"
          },
          "liens": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LienDeFicheDto"
            }
          }
        },
        "required": [
          "id",
          "libelle",
          "ordre",
          "compteur",
          "liens"
        ]
      },
      "FicheEntiteDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "typeEntiteId": {
            "type": "string",
            "format": "uuid"
          },
          "typeCode": {
            "type": "string"
          },
          "libelle": {
            "type": "string"
          },
          "visibilite": {
            "enum": [
              "public",
              "restreint",
              "prive"
            ],
            "type": "string"
          },
          "etat": {
            "enum": [
              "actif",
              "archive"
            ],
            "type": "string"
          },
          "modifieLe": {
            "type": "string",
            "format": "date-time"
          },
          "typeLibelle": {
            "type": "string"
          },
          "valeurs": {
            "type": "object",
            "additionalProperties": true,
            "description": "Projection des faits **visibles par cet agent** — recomposée à la lecture, et non recopiée depuis la colonne, qui ignore la visibilité"
          },
          "contenuLisible": {
            "type": "boolean",
            "description": "Faux sur une entité restreinte non habilitée : l’objet est visible, son contenu non"
          },
          "note": {
            "type": "string",
            "nullable": true
          },
          "dossiers": {
            "description": "Dossiers qui suivent cette entité, filtrés. Une entité peut appartenir à plusieurs dossiers, et la fiche l’indique.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RattachementDto"
            }
          },
          "champs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChampDeFicheDto"
            }
          },
          "onglets": {
            "description": "Onglets configurés pour ce type d’entité, déjà peuplés de leurs liens",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OngletPeupleDto"
            }
          },
          "liensHorsOnglet": {
            "description": "Liens qu’aucun onglet ne regroupe — ils resteraient invisibles sinon, ce qui ferait passer une erreur de configuration pour une absence de donnée",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LienDeFicheDto"
            }
          },
          "liens": {
            "description": "Tous les liens visibles, à plat",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LienDeFicheDto"
            }
          },
          "creeLe": {
            "type": "string",
            "format": "date-time"
          },
          "fusionneeVersId": {
            "type": "string",
            "nullable": true,
            "format": "uuid"
          }
        },
        "required": [
          "id",
          "typeEntiteId",
          "typeCode",
          "libelle",
          "visibilite",
          "etat",
          "modifieLe",
          "typeLibelle",
          "valeurs",
          "contenuLisible",
          "note",
          "dossiers",
          "champs",
          "onglets",
          "liensHorsOnglet",
          "liens",
          "creeLe",
          "fusionneeVersId"
        ]
      },
      "EvenementHistoriqueDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "nature": {
            "enum": [
              "fait",
              "modification"
            ],
            "type": "string",
            "description": "Un fait sorti du graphe actif, ou une trace d’écriture du journal d’audit"
          },
          "libelle": {
            "type": "string"
          },
          "source": {
            "type": "string",
            "nullable": true
          },
          "fiabilite": {
            "type": "number",
            "nullable": true
          },
          "auteur": {
            "type": "string",
            "nullable": true,
            "description": "« agent supprimé » lorsque le compte a été anonymisé"
          },
          "survenuLe": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "nature",
          "libelle",
          "source",
          "fiabilite",
          "auteur",
          "survenuLe"
        ]
      },
      "ChampSaisiDto": {
        "type": "object",
        "properties": {
          "source": {
            "type": "string",
            "minLength": 1,
            "maxLength": 512,
            "example": "Rapport d'intervention n°2291"
          },
          "fiabilite": {
            "type": "number",
            "minimum": 1,
            "maximum": 4
          },
          "dateConstatation": {
            "type": "string",
            "example": "2026-08-07",
            "format": "date"
          },
          "visibilite": {
            "enum": [
              "public",
              "restreint",
              "prive"
            ],
            "type": "string"
          },
          "definitionChampId": {
            "type": "string",
            "format": "uuid"
          },
          "valeur": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "number"
              },
              {
                "type": "boolean"
              }
            ],
            "description": "Texte, nombre, booléen ou valeur de liste, selon le type du champ"
          }
        },
        "required": [
          "definitionChampId",
          "valeur"
        ]
      },
      "LienSaisiDto": {
        "type": "object",
        "properties": {
          "source": {
            "type": "string",
            "minLength": 1,
            "maxLength": 512,
            "example": "Rapport d'intervention n°2291"
          },
          "fiabilite": {
            "type": "number",
            "minimum": 1,
            "maximum": 4
          },
          "dateConstatation": {
            "type": "string",
            "example": "2026-08-07",
            "format": "date"
          },
          "visibilite": {
            "enum": [
              "public",
              "restreint",
              "prive"
            ],
            "type": "string"
          },
          "typeLienId": {
            "type": "string",
            "format": "uuid"
          },
          "cibleId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "typeLienId",
          "cibleId"
        ]
      },
      "CreationEntiteDto": {
        "type": "object",
        "properties": {
          "source": {
            "type": "string",
            "minLength": 1,
            "maxLength": 512,
            "example": "Rapport d'intervention n°2291"
          },
          "fiabilite": {
            "type": "number",
            "minimum": 1,
            "maximum": 4
          },
          "dateConstatation": {
            "type": "string",
            "example": "2026-08-07",
            "format": "date"
          },
          "visibilite": {
            "enum": [
              "public",
              "restreint",
              "prive"
            ],
            "type": "string"
          },
          "typeEntiteId": {
            "type": "string",
            "format": "uuid"
          },
          "dossierId": {
            "type": "string",
            "format": "uuid",
            "description": "Dossier depuis lequel la saisie a lieu. Ses faits en héritent la visibilité ; l’entité, elle, ne porte que la sienne."
          },
          "note": {
            "type": "string",
            "maxLength": 10000,
            "description": "Champ libre, sans source ni fiabilité — ce n’est pas un fait"
          },
          "champs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChampSaisiDto"
            }
          },
          "liens": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LienSaisiDto"
            }
          }
        },
        "required": [
          "typeEntiteId"
        ]
      },
      "ModificationEntiteDto": {
        "type": "object",
        "properties": {
          "note": {
            "type": "string",
            "maxLength": 10000
          },
          "visibilite": {
            "enum": [
              "public",
              "restreint",
              "prive"
            ],
            "type": "string",
            "description": "Exige la permission visibilite.definir"
          }
        }
      },
      "FusionDto": {
        "type": "object",
        "properties": {
          "versId": {
            "type": "string",
            "format": "uuid",
            "description": "La fiche qui subsiste"
          }
        },
        "required": [
          "versId"
        ]
      },
...
    }
  }
}